+++ Playground Panic/Project/Game/ProjectileObject.h
@@ -15,7 +15,7 @@ class ProjectileObject : public GameObject
 {
 public:
 	ProjectileObject(sf::Sprite* sprite, float radius, Collider* collider = nullptr);
-	void Initialize(sf::Vector2f origin, sf::Vector2i target);
+	void Initialize(sf::Vector2f origin, sf::Vector2f target);
 
 	void Update(float deltatime, float global_speed, sf::Vector2i mousePos);
 	void Collision(bool hit);
@@ -30,7 +30,7 @@ protected:
 	/*sf::Vector2f m_position;
 	sf::Vector2f m_target;*/
 	sf::Vector2f m_velocity;
-	sf::Vector2i m_mouse_position;
+	sf::Vector2f m_mouse_position;
 
 	AnimatedSprite *m_current_animation;
 	std::map<std::string, AnimatedSprite*> m_animations;
